+++ Changes.md
@@ -322,7 +322,12 @@ Build
 - TBB : Updated to version 2021.13.0.
 - USD : Updated to version 26.05.
 
-1.6.21.x (relative to 1.6.21.5)
+1.6.21.x (relative to 1.6.21.6)
+========
+
+
+
+1.6.21.6 (relative to 1.6.21.5)
 ========
 
 Fixes